Server imports, added sending client event to client network manager
