+++ package.json
@@ -23,6 +23,7 @@
   },
   "devDependencies": {
     "@babel/core": "^7.10.2",
+    "@babel/register": "^7.10.1",
     "concurrently": "^5.2.0",
     "cssnano": "^4.1.10"
   }
